move ScriptWriter in ZEditor
